+++ raytracer/include/shape.hh
@@ -13,7 +13,7 @@ public:
   Shape(const Vector3& pos, const Attributes& attr, const Color& color);
   virtual Vector3 intersect(const Ray& ray) = 0;
   virtual Color& apply_point_lights(const Input& file, Color& r,
-                                    const Vector3& intersect, int ttl) = 0;
+                                    const Vector3& intersect, size_t ttl) = 0;
   virtual Color& apply_ambiant_light(const Input& file, Color& r) = 0;
   virtual Color& apply_directional_lights(const Input& file, Color& r,
                                           const Vector3& intersect) = 0;
